Key store2.0

+++ Quantumnetwork Controllcenter/src/KeyStore/KeyStoreDbManager.java
@@ -6,11 +6,20 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
+import java.util.AbstractMap;
 import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+/**
+ * This class supplies methods for creating, editing, getting and deleting entries from the KeyStore.db which holds all the keys
+ * @author Aron Hernandez
+ */
 
 public class KeyStoreDbManager {
     private static final String dataBaseName = "KeyStore.db";
-    private static final String tableName = "KeyInformations";
+    private static final String tableName = "KeyStorage";
 
 
     /**
@@ -30,8 +39,9 @@ public class KeyStoreDbManager {
 
         } catch (ClassNotFoundException | SQLException e) {
 
-            System.out.println("Connection to database failed!" + "\n");
-            e.printStackTrace();
+            System.err.println("Connection to database failed!" + "\n");
+            System.err.println(e.toString());
+
         }
 
         return con;
@@ -42,7 +52,7 @@ public class KeyStoreDbManager {
      *
      * @returns True if Database and table were created successfully, False otherwise
      */
-     static boolean createNewKeyStoreAndTable() {
+    protected static boolean createNewKeyStoreAndTable() {
 
         try {
             Connection conn = KeyStoreDbManager.connect();
@@ -50,29 +60,30 @@ public class KeyStoreDbManager {
 
             System.out.println("Database was created successfully!");
 
-                // create Table
-            String sql = "CREATE TABLE IF NOT EXISTS " + tableName +
-                        " (KeyStreamId CHAR(128) ," +
-                        " KeyBuffer INTEGER NOT NULL, " + // Index muss definitiv noch angepasst werden, Frage im Treffen heute nach
-                        " Index_ INTEGER NOT NULL , " +
-                        " Source_ TEXT NOT NULL, " +
-                        " Destination TEXT NOT NULL, " +
-                        "PRIMARY KEY (KeyStreamId, Index_))";
+            // create Table
+            String keyInformationSQL = "CREATE TABLE IF NOT EXISTS " + tableName +
+                    " (KeyStreamId CHAR(128) UNIQUE ," +
+                    " KeyBuffer INTEGER UNIQUE , " + // Index muss definitiv noch angepasst werden, Frage im Treffen heute nach
+                    " Index_ INTEGER NOT NULL , " +
+                    " Source_ TEXT NOT NULL, " +
+                    " Destination TEXT NOT NULL, " +
+                    " Used BOOLEAN NOT NULL, " +
+                    "PRIMARY KEY (KeyStreamId, Index_))";
 
-            stmnt.executeUpdate(sql);
+            stmnt.executeUpdate(keyInformationSQL);
             System.out.println("Table creation successful!");
+
             stmnt.close();
             conn.close();
 
             return true;
 
         } catch (SQLException e) {
-            System.out.println("Database creation failed!" + "\n");
-            e.printStackTrace();
+            System.err.println("Database creation failed!" + "\n");
+            System.err.println(e.toString());
 
             return false;
         }
-
     }
 
      /*
@@ -80,31 +91,38 @@ public class KeyStoreDbManager {
       * Why is the javadoc incomplete, and why does it call it an array of char bytes? Char is 2 bytes, not 1 byte.
       */
 
-    /** Inserts a new Entry to the DB
+
+    /**Following the Etsi paper there should be 2 Tables. #1 which holds all the information about a key
+     * and #2 which stores the real keys.
+     *
+     * ---> Method inserts a new Entry in table #1 (KeyInformation)
      *
      * @param keyStreamID reference ID to locate a key
      * @param keyBuffer array of char bytes ()
      * @param index index for the new Key
      * @param source identifier for the source application
      * @param destination identifier for the destination application
+     * @param used boolean parameter signaling whether a key has been used already
      */
-    public static boolean insertToDb(  String keyStreamID, int keyBuffer, int index, String source, String destination ){
+
+    public static boolean insertToKeyStore(String keyStreamID, byte[] keyBuffer, int index, String source, String destination, boolean used ){
 
         try{
             Connection conn = connect();
 
-            String sql = "INSERT INTO KeyInformations(KeyStreamID, KeyBuffer, Index_, Source_, Destination)  VALUES(?,?,?,?,?)";
+            String sql = "INSERT INTO " + tableName + "(KeyStreamID, KeyBuffer, Index_, Source_, Destination, Used)  VALUES(?,?,?,?,?,?)";
 
             PreparedStatement prepStmnt = conn.prepareStatement(sql);
 
             prepStmnt.setString(1, keyStreamID);
-            prepStmnt.setInt(2, keyBuffer);
+            prepStmnt.setBytes(2, keyBuffer);
             prepStmnt.setInt(3, index);
             prepStmnt.setString(4, source);
             prepStmnt.setString(5, destination);
+            prepStmnt.setBoolean(6, used);
 
             prepStmnt.executeUpdate();
-            System.out.println("Insertion to DB was successful");
+            System.out.println("Insertion to KeyInformation table was successful");
 
             prepStmnt.close();
             conn.close();
@@ -113,44 +131,59 @@ public class KeyStoreDbManager {
         }
 
         catch (SQLException e ){
-            System.out.println("Insertion to DB failed!" + "\n" );
-            e.printStackTrace();
+            System.err.println("Insertion to KeyInformation table failed!" + "\n");
+            System.err.println(e.toString());
+
         }
         return false;
     }
 
-    /** (Api needs a function which reserves an empty KeyStreamID, therefore we should be able to rename one by the index (i guess))
+    /**
      *
-     * @param index the index of the keyStreamID to be renamed
-     * @param newID new KeyStreamID
-     * @return True if operation was successful, False otherwise
+     * @param keyStreamID reference ID to locate a Key.
+     * @param key the new Key as a byte[]
+     * @return True if operation was successful, false otherwise.
      */
-    public static boolean updateKeyStreamID (int index, String newID) {
-        try {
-            Connection conn = connect();
-            String sql = "UPDATE " + tableName + " SET KeyStreamId = ? WHERE Index_ = ?";
-            PreparedStatement pstmnt = conn.prepareStatement(sql);
-            pstmnt.setString(1, newID);
-            pstmnt.setInt(2, index);
-            pstmnt.executeUpdate();
+    protected static boolean addKeyBuffer(String keyStreamID, byte[] key){
 
-            pstmnt.close();
-            conn.close();
-            return true;
+        if (!doesKeyStreamIdExist(keyStreamID)){
+            System.err.println("There is no Entry with this KeyStreamID" + "\n");
+            return false;
+        }
+        KeyStoreObject obj = getEntryFromKeyStore(keyStreamID);
 
-        } catch (Exception e) {
-            System.err.println("Renaming entry failed!" + "\n");
-            e.printStackTrace();
+        if (obj.getKeyBuffer() == null) {
+            try {
+                Connection conn = connect();
+                String sql = "UPDATE " + tableName + " SET KeyBuffer = ? WHERE KeyStreamID = ?";
+                PreparedStatement pstmnt = conn.prepareStatement(sql);
+                pstmnt.setBytes(1, key);
+                pstmnt.setString(2, keyStreamID);
+                pstmnt.executeUpdate();
+
+                pstmnt.close();
+                conn.close();
+                System.out.println("Renaming entry in KeyInformation table succeeded!\" + \"\\n\"");
+                return true;
+
+            } catch (SQLException e) {
+                System.err.println("Renaming entry in KeyInformation table failed!" + "\n");
+                System.err.println(e.toString());
+
+                return false;
+            }
+        }
+        else {
+            System.err.println("This Entry already has a dedicated keyBuffer." + "\n");
             return false;
         }
     }
-
 
     /** Displays all entries on the console
      *
      * @return True if output was displayed correctly, False otherwise
      */
-    static boolean selectAll(){
+    protected static boolean selectAll(){
 
         try {
             String sql = "SELECT * FROM " + tableName;
@@ -170,8 +203,8 @@ public class KeyStoreDbManager {
             conn.close();
             return true;
         } catch (SQLException e) {
-            System.out.println( "Selecting every entry from KeyStore.db failed!" );
-            e.printStackTrace();
+            System.err.println( "Selecting every entry from KeyStore.db failed!" );
+            System.err.println(e.toString());
             return false;
         }
     }
@@ -181,7 +214,11 @@ public class KeyStoreDbManager {
      * @param keyStreamID the ID of the key that needs to be deleted from the DB
      * @return True if operation was successful, False otherwise
      */
-    static boolean deleteEntryByID(String keyStreamID){
+    protected static boolean deleteKeyInformationByID(String keyStreamID){
+        if (!KeyStoreDbManager.doesKeyStreamIdExist(keyStreamID)){
+            System.err.println("Deleting Key entry from DB failed because the given keyStreamID does not exist!"  + "\n");
+            return false;
+        }
 
         try {
             String sql = "DELETE FROM " + tableName + " WHERE KeyStreamId= ?";
@@ -191,7 +228,7 @@ public class KeyStoreDbManager {
 
             pstmnt.setString(1, keyStreamID);
             pstmnt.executeUpdate();
-            System.out.println("Entry was deleted successfully!");
+            System.out.println("Entry from KeyInformation table was deleted successfully!");
             pstmnt.close();
             conn.close();
 
@@ -199,19 +236,51 @@ public class KeyStoreDbManager {
         }
 
         catch (SQLException e) {
-            System.out.println("Entry deletion failed!" + "\n");
-            e.printStackTrace();
+            System.err.println("Deleting entry from KeyInformation table failed!" + "\n");
+            System.err.println(e.toString());
             return false;
         }
 
     }
 
-    /** Get a new KeystoreObject by the corresponding keyStreamID
+    /** Change the "Used" parameter of a KeyStore Entry from unused(=0) to used(=1)
+     *
+     * @param keyStreamID reference ID for a Key
+     * @return true if operation succeeded, false otherwise
+     */
+    protected static boolean changeKeyToUsed(String keyStreamID){
+        try {
+            Connection conn = connect();
+
+            String sql = "UPDATE " + tableName + " SET Used = 1 WHERE KeyStreamID = ?";
+            PreparedStatement pstmnt = conn.prepareStatement(sql);
+
+            pstmnt.setString(1, keyStreamID);
+            pstmnt.executeUpdate();
+            System.out.println("Changed key from Unused to Used");
+            pstmnt.close();
+            conn.close();
+
+            return true;
+
+        } catch (SQLException e) {
+            System.err.println("Changing the used Parameter of Entry failed " + "\n");
+            System.err.println(e.toString());
+            return false;
+        }
+    }
+
+    /** Get a new KeyInformationObject by the corresponding keyStreamID
      *
      * @param keyStreamID identifier of the key that needs to be wrapped in a KeyStoreObject
-     * @return a new KeyStoreObject containing all the KeyInformations from the Entry with corresponding KeyStreamId
+     * @return a new KeyInformationObject containing all the KeyInformation from the Entry with corresponding KeyStreamId
      */
-    public static KeyStoreObject getEntry(String keyStreamID) {
+     protected static KeyStoreObject getEntryFromKeyStore(String keyStreamID) {
+         if (!doesKeyStreamIdExist(keyStreamID)){
+             System.err.println("There is no Entry with this KeyStreamID!"  + "\n");
+             return null;
+         }
+
         try {
             Connection conn = connect();
 
@@ -222,26 +291,27 @@ public class KeyStoreDbManager {
             ResultSet rs = stmnt.executeQuery(sql);
 
             KeyStoreObject object = new KeyStoreObject(rs.getString("KeyStreamId"),
-                    rs.getInt("KeyBuffer"), rs.getInt("Index_"),
-                    rs.getString("Source_"), rs.getString("Destination"));
+                    rs.getBytes("KeyBuffer"), rs.getInt("Index_"),
+                    rs.getString("Source_"), rs.getString("Destination"), rs.getBoolean("Used"));
 
             stmnt.close();
             conn.close();
+            System.out.println("Selecting entry from KeyInformation table was successful!" + "\n");
             return object;
 
 
         } catch (SQLException e) {
-            System.out.println("Selecting the entry failed!" + "\n");
-            e.printStackTrace();
+            System.err.println("Selecting entry from KeyInformation table failed!" + "\n");
+            System.err.println(e.toString());
             return null;
         }
     }
 
-    /** Stores all the Entrys of the DB in an ArrayList and returns the List
+    /** Stores all the Entries of the KeyInformation table in an ArrayList and returns the List
      *
-     * @return a ArrayList of KeyStoreObjects which contain information about the keys currently in storage
+     * @return a ArrayList of KeyInformationObject which contain information about the keys currently in storage
      */
-    public static ArrayList<KeyStoreObject> getEntriesAsList () {
+     protected static ArrayList<KeyStoreObject> getKeyStoreAsList() {
         try {
             Connection conn = connect();
 
@@ -251,18 +321,127 @@ public class KeyStoreDbManager {
             ArrayList<KeyStoreObject> result = new ArrayList<>();
             while(rs.next()) {
                 KeyStoreObject res = new KeyStoreObject(rs.getString("KeyStreamID"),
-                        rs.getInt("KeyBuffer"), rs.getInt("Index_"), rs.getString("Source_"), rs.getString("Destination"));
+                        rs.getBytes("KeyBuffer"), rs.getInt("Index_"),
+                        rs.getString("Source_"), rs.getString("Destination"), rs.getBoolean("Used"));
                 result.add(res);
             }
             stmnt.close();
             conn.close();
+            System.out.println("Generating list of KeyInformation table entries was successful" + "\n");
             return result;
-        } catch (Exception e) {
-            System.out.println("Generating a list of all entries failed!" + "\n");
-            e.printStackTrace();
+        } catch (SQLException e) {
+            System.err.println("Generating list of all entries from KeyInformation table failed!" + "\n");
+            System.err.println(e.toString());
             return null;
         }
     }
+
+    /** Check whether this keyStreamId exists in the KeyStore or not.
+     *
+     * @param keyStreamID reference ID to locate a key
+     * @return true if the keyStreamID exists, false otherwise
+     */
+   protected static boolean doesKeyStreamIdExist(String keyStreamID){
+        List<String> keyIdList = KeyStoreDbManager.getKeyStoreAsList().stream().map(obj -> new String(obj.getID())).collect(Collectors.toList());
+
+        return keyIdList.contains(keyStreamID);
+
+    }
+
+    /*
+     *  ------ API Methods ------
+     */
+
+    /** from etsi paper: Reserve an association (Key_stream_ID) for a set of future keys at both ends of the QKD link
+     through this distributed QKD key management layer and establish a set of parameters that
+     define the expected levels of key service. This function shall block until the peers are
+     connected or until the Timeout value in the QoS parameter is exceeded.
+     *
+     * @param source identifier for the source application
+     * @param destination identifier for the destination application
+     * @param qosParameters multiple parameters that are delivered as a QoS Object                                       //Weiß selber ned wie der Parameter aussehen soll
+     * @param keyStreamID reference ID to locate a key
+     * @return number of status
+     */
+    static int open_Connect(String source, String destination, QoS qosParameters, String keyStreamID, boolean peerConnected){
+
+        int timeout = qosParameters.getTimeout();
+        int exceeded = 10;// bin mir nicht sicher ob das überhaupt in die methode gehört oder ob das von woanders angesteuert wird
+
+        //return 5 -----> if KeyStramID already exists
+        if(!doesKeyStreamIdExist(keyStreamID)){
+            return 5;
+        }
+
+        boolean insertbool = insertToKeyStore(keyStreamID, null, 0, source, destination, false);
+
+        if(insertbool) {
+            //return 0 -----> if everything was succesfull
+            if (peerConnected) return 0;
+
+            //return 1 -----> if connection was established but peer is not connected
+            return 1;
+
+        }
+        else{
+            return -1; // indicating that something went wrong
+        }
+
+        //TODO
+
+        //return 6 -----> if the timeout parameter is exceeded -----> when is it exceeded tho?
+
+        //return 7 -----> if OPEN failed because requested QoS settings could not be met, counter proposal included in return has occurred
+
+    }
+
+    /** from etsi: Terminate the association established for this Key_stream_ID. No further keys shall be
+     allocated for this Key_stream_ID after the association has been closed. Due to timing
+     differences at the other end of the link this peer operation will happen at some other time and
+     any unused keys shall be held until that occurs and then discarded or the TTL (Time To Live
+     QoS parameter) has expired.
+     *
+     * @param keyStreamID reference ID to locate a key
+     * @return number of status
+     */
+    static int close(String keyStreamID, boolean peerConnected){
+        //delete keyInformation for this ID
+
+        //return 0 -----> if everything was succesfull
+        if (doesKeyStreamIdExist(keyStreamID)){
+           deleteKeyInformationByID(keyStreamID);
+           if (peerConnected) return 0;
+            return 1;
+        }
+        System.err.println("There is no Entry with this KeyStreamID" + "\n");
+        return -1; // -1 = new Status parameter indicating that the operation failed
+
+
+    }
+
+    /** Etsi paper page 9 latest version for more information
+     *
+     * @param keyStreamID reference ID to locate a key
+     * @return number of status + the key
+     */
+    static Map.Entry<byte[], Integer> get_Key(String keyStreamID) {
+        AbstractMap.SimpleEntry<byte[], Integer> pair;
+
+        if(doesKeyStreamIdExist(keyStreamID)){
+            KeyStoreObject obj = getEntryFromKeyStore(keyStreamID);
+            int index = obj.getIndex();
+            byte[] key = obj.getKeyBuffer();
+
+            return new AbstractMap.SimpleEntry<byte[], Integer>(key, index);
+        }
+
+        return null;
+
+
+    }
+
+
+
